mailboxes/requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.16.0
cryptography==42.0.8
dnspython==2.6.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
imap-tools==1.6.0
magic-filter==1.0.12
motor==3.5.1
multidict==6.0.5
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pymongo==4.8.0
python-dotenv==1.0.1
soupsieve==2.5
SQLAlchemy==2.0.31
typing_extensions==4.12.2
yarl==1.9.4
